ObjectDetection_app-main/requirements.txt

absl-py==1.0.0
altgraph==0.17.2
amqp==5.0.9
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.2.10
astor==0.8.1
async-generator==1.10
billiard==3.6.4.0
ca-certificates==2021.10.8
cached-property==1.5.2
celery==5.1.2==
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.11
click==7.1.2==
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
contextlib2==21.6.0
cycler==0.11.0
cython==0.29.27
dataclasses==0.8
deprecated==1.2.13
django==3.1.2
django-celery-results==2.2.0
django-cors-headers==3.10.1
django-mssql-backend==2.8.1
djangorestframework==3.13.1
entrypoints==0.3
flask==2.0.2
flatbuffers==1.12
freetype==2.10.4
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.43.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.3
importlib_metadata==3.10.1
imutils==0.5.4
ipykernel==5.5.6
ipython==7.16.3
itsdangerous==2.0.1
jedi==0.17.2
jinja2==3.0.3
jpeg==9b
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0==
jupyter-core==4.9.1==
jupyterlab-pygments==0.1.2==
jupyterlab-widgets==1.0.2==
keras-applications==1.0.8==
keras-preprocessing==1.1.2==
kiwisolver==1.3.1==
kombu==5.1.0==
libpng==1.6.37==h2a8f88b_0
libpq==12.2==hb652d5d_1
libprotobuf==3.13.0.1==h200bbdf_0
libtiff==4.1.0==h56a325e_1
lxml==4.7.1==
lz4-c==1.9.2==hf4a77e7_3
markdown==3.3.6==
matplotlib==3.3.4==
more-itertools==8.12.0==pyhd8ed1ab_0
mssql-django==1.1.2==
nbclient==0.5.9==
nbconvert==6.0.7==
nest-asyncio==1.5.4==
notebook==6.4.8==
numpy==1.19.5==
olefile==0.46==py36_0
onnx==1.10.2==
onnxruntime-gpu==1.9.0==
opencv-contrib-python==4.5.5.62==
opencv-python==4.5.5.62==
openssl==3.0.0==h8ffe710_2
opt-einsum==3.3.0==
packaging==21.3==
parso==0.7.1==
pefile==2021.9.3==
pillow==8.4.0==
pip==21.3.1==
prometheus-client==0.13.1==
prompt-toolkit==3.0.26==
protobuf==3.19.4==
psycopg2==2.8.6==py36hcd4344a_1
pygments==2.11.2==
pyinstaller==4.9==
pyinstaller-hooks-contrib==2022.0==
pyodbc==4.0.32==
pyparsing==3.0.7==
pyqt5==5.15.6==
pyqt5-qt5==5.15.2==
pyqt5-sip==12.9.1==
pyqt6==6.2.3==
pyqt6-qt6==6.2.3==
pyqt6-sip==13.2.1==
pyrsistent==0.18.0==
python==3.6.13==h3758d61_0
python-dateutil==2.8.2==
python-snap7==1.1==
python_abi==3.6==2_cp36m
pytz==2021.3==
pywin32==303==
pywin32-ctypes==0.2.0==
pywinpty==1.1.6==
pyzmq==22.3.0==
qdarkstyle==3.0.3==
qtconsole==5.2.2==
qtpy==2.0.0==
qtwidgets==0.18==
redis==4.1.3==
requests==2.27.1==
rq==1.10.1==
setuptools==59.6.0==
shapely==1.8.0==
six==1.16.0==
sqlalchemy==1.4.13==py36h68aa20f_0
sqlite==3.37.2==h2bbff1b_0
sqlparse==0.4.1==py_0
tensorboard==1.15.0==
tensorflow-estimator==1.15.1==
tensorflow-gpu==1.15.0==
termcolor==1.1.0==
terminado==0.12.1==
tf2onnx==1.9.3==
tk==8.6.10==he774522_0
torch==1.3.0+cpu==
torchvision==0.4.1+cpu==
typing-extensions==4.0.1==
typing_extensions==3.10.0.2==pyh06a4308_0
urllib3==1.26.8==
vc==14.2==h21ff451_1
vine==5.0.0==pyhd8ed1ab_1
vs2015_runtime==14.27.29016==h5e58377_2
wcwidth==0.2.5==
webencodings==0.5.1==
werkzeug==2.0.2==
wheel==0.37.1==
widgetsnbextension==3.5.2==
wincertstore==0.2==py36h7fe50ca_0
wrapt==1.13.3==
xz==5.2.5==h62dcd97_0
zipp==3.6.0
zlib==1.2.11
zstd==1.4.4